+++ docs/CHANGELOG.md
@@ -13,6 +13,11 @@ The versioning system is basically the following:
   - If there's a really "big" change.
 - As mentioned, the "major" version is 0.
 
+## main
+
+- Warn when a `.sig` file doesn't contain 1 `signature`.
+- Warn when a `.fun` file doesn't contain 1 `functor`.
+
 ## v0.13.5
 
 - Tweak inlay hint parentheses around `as` patterns.